Update reverse_proxy.md
Documantion says: "By default, Caddy passes thru incoming headers [...], with two exceptions:". But technically there is another another header that is added (if not present): `Accept-Encoding`. 

When the client is a web browser this header should be already set. But for curl or similar tools, that might not be the case.