docs: Elaborate on how the `php_fastcgi` directive works
This is mostly taken from a post I wrote on the forums last month https://caddy.community/t/php-fastcgi-only-serving-php-files-or-wildcard-possible/13918/3 which I thought was a good and thorough explanation of how the directive works, and why it works that way. Did a bit of polish on the wording the fit the language the docs usually use (IMO anyways).

Also added a ToC since the page gets a bit long after these additions, best to give users a way to jump down to the examples quickly.